Rename `AtLeast` and `OneOf` rules
