+++ src/RuleInterface.php
@@ -15,7 +15,7 @@ interface RuleInterface
      * A matching handler name or an instance used for processing this rule.
      *
      * While not required, for naming of handlers' classes it's recommended to use a rule class name with "Handler"
-     * suffix, so for `AtLeast` rule class name the handler class name will be `AtLeastHandler` and so on.
+     * suffix, so for `FilledAtLeast` rule class name the handler class name will be `FilledAtLeastHandler` and so on.
      *
      * All packages handlers are stored within the same namespace as rules, but this is not a strict requirement.
      *
