+++ src/RulesProvider/AttributesRulesProvider.php
@@ -126,7 +126,7 @@ final class AttributesRulesProvider implements RulesProviderInterface
      *
      * ```php
      * [
-     *     [new AtLeast(['name', 'author'])], // Rules not bound to a specific property.
+     *     [new FilledAtLeast(['name', 'author'])], // Rules not bound to a specific property.
      *     'files' => [new Count(max: 3)], // Property specific rules.
      * ],
      * ```
